Move the PostgisGeometryDecoder into the util package
