+++ core/src/main/java/org/apache/calcite/util/BuiltInMethod.java
@@ -122,7 +122,6 @@ import org.apache.calcite.sql.SqlJsonExistsErrorBehavior;
 import org.apache.calcite.sql.SqlJsonQueryEmptyOrErrorBehavior;
 import org.apache.calcite.sql.SqlJsonQueryWrapperBehavior;
 import org.apache.calcite.sql.SqlJsonValueEmptyOrErrorBehavior;
-import org.apache.calcite.sql.dialect.PostgisGeometryDecoder;
 
 import com.google.common.collect.ImmutableMap;
